Externalize all queries from TagList, QuestionList and SessionList
